// src/app/interfaces/linea-etica.ts
export interface LineaEtica {
  formId: number;
  strangeness: string;
  strangenessDate: string;
  thePlaceOfScene: string;
  // knowTheNameOfInvolvedPeople: string;
   // nameOfInvolvedPeople: string;
  strangenessTime: number;
  companyKnows: string;
  peopleNameInCompany: string;
  fileIds: Array<number>;
  wantRegisterData:  string;
  customerName1: string;
  customerName2: string;
  customerlastName1: string;
  customerlastName2: string;
  phoneNumber: number;
  interestGroupId: number;
  otherGroup: string;
  email: string;
  timeUnitId: number;
  politicId: number;
  recaptchaToken: string;
}